+++ contrib/common.cpp
@@ -1,5 +1,5 @@
 #include "common.h"
-#include "../src/cufinufft.h"
+#include <cufinufft.h>
 #include <math.h>
 #include <stdio.h>
 #include <vector>
