+++ include/memtransfer.h
@@ -1,7 +1,7 @@
 #ifndef __MEMTRANSFER_H__
 #define __MEMTRANSFER_H__
 
-#include "cufinufft.h"
+#include <cufinufft.h>
 
 int allocgpumem1d_plan(cufinufft_plan *d_plan);
 int allocgpumem1d_nupts(cufinufft_plan *d_plan);
