+++ src/2d/interp2d_wrapper.cu
@@ -7,9 +7,9 @@
 #include <cub/device/device_scan.cuh>
 
 #include <cuComplex.h>
-#include "../cuspreadinterp.h"
-#include "../memtransfer.h"
-#include "../profile.h"
+#include <cuspreadinterp.h>
+#include <memtransfer.h>
+#include <profile.h>
 
 using namespace std;
