+++ src/2d/spread2d_wrapper_paul.cu
@@ -8,8 +8,8 @@
 #include <cub/device/device_scan.cuh>
 
 #include <cuComplex.h>
-#include "../cuspreadinterp.h"
-#include "../memtransfer.h"
+#include <cuspreadinterp.h>
+#include <memtransfer.h>
 
 using namespace std;
