+++ src/README
@@ -1,20 +1,13 @@
 This folder contains the main source files of the GPU implementations. 
-- cufinufft.h
 - cufinufft.cu
   Five main stages of cufinufft API. 
   (1) cufinufft_default_opts, (2) cufinufft_makeplan, (3) cufinufft_setNUpts, (3) cufinufft_exec, (3) cufinufft_destroy.
 
-- memtransfer.h
 - memtransfer_wrapper.cu
   Wrapper of allocation and free GPU memories for different dimensions and methods.
 
-- deconvolve.h
 - deconvolve_wrapper.cu
   GPU kernels and wrappers of deconvolve and amplify the input/output coefficients by correction factor. (Step 3 in Type 1; Step 1 in Type 2)
 
-- spreadinterp.h
-  Definitions of GPU kernels and wrappers for spreader and interpolator for 2,3 dimensions under 2d and 3d folders.
-
-- profile.h
 - profile.cu
   Codes for using NVProf
