+++ src/cufinufft.cu
@@ -5,10 +5,10 @@
 #include <complex>
 #include <cufft.h>
 
-#include "cufinufft.h"
-#include "cuspreadinterp.h"
-#include "cudeconvolve.h"
-#include "memtransfer.h"
+#include <cufinufft.h>
+#include <cuspreadinterp.h>
+#include <cudeconvolve.h>
+#include <memtransfer.h>
 
 using namespace std;
