+++ src/profile.cu
@@ -1,4 +1,4 @@
-#include "profile.h"
+#include <profile.h>
 #include <nvToolsExt.h>
 #include <cstdio>
