+++ test/cufinufft2d1_test.cu
@@ -4,7 +4,7 @@
 #include <helper_cuda.h>
 #include <complex>
 
-#include "../src/cufinufft.h"
+#include <cufinufft.h>
 #include "../contrib/utils.h"
 
 using namespace std;
