+++ test/spread_3d.cu
@@ -3,8 +3,7 @@
 #include <math.h>
 #include <helper_cuda.h>
 #include <complex>
-#include "../src/cufinufft.h"
-#include "../src/cuspreadinterp.h"
+#include <cuspreadinterp.h>
 #include "../contrib/utils.h"
 
 using namespace std;
